Single Controller Mode
Now, by pressing the Y button on the controller or button 5 on the joystick, Single Controller mode should be toggleable. This will allow the joystick to operate by itself without the gamepad. Also, using the rope motors individually instead of in unison is commented out, as well as the ball release mechanism code.